feat: create a secure login endpoint with token revocation

+++ internals/pkg/repository/repository.go
@@ -5,7 +5,7 @@ import "github.com/ishanshre/GoFileServerAPI/internals/pkg/models"
 type Repository interface {
 
 	//user interface
-	GetUsers(limit, page int64) ([]*models.User, error)
+	GetUsers(limit, page int) ([]*models.User, error)
 	GetUserByUsername(username string) (*models.User, error)
 	GetUserByEmail(email string) (*models.User, error)
 	UsernameExists(username string) error
